refactor(workstation): extract shared types + visual helpers from inkRuntime (#890 phase 5a.0)
Foundation for the inkRuntime.ts split tracked in phase 5a of #890.
The 6,039-LOC file decomposes into roughly three layers:

  1. Imports + small helpers + startInkInteractiveLog + loadInkRuntime
     (lines 1-843) — orchestration, stays in commands/log/
  2. LogInkApp React component (lines 844-2780) — phase 5b target
  3. ~50 render*Surface and chrome render helpers (lines 2781+) —
     phase 5a target

Phase 5a.1+ will extract layer 3 into workstation/surfaces/<view>/ and
workstation/runtime/. Each surface module needs the same set of shared
types (LogInkContext, LogInkComponents, LogInkRuntime, LogInkComponentDeps)
and visual helpers (compactHash, focusBorderColor, panelTitle,
diffLineProps, statusCodeColor, formatChangedFileStats, sidebarTabLabel).

This PR (5a.0) extracts that shared set so subsequent surface PRs become
mechanical pure cuts:

  - New: src/workstation/runtime/types.ts (109 LOC)
  - New: src/workstation/runtime/utils.ts (144 LOC)
  - inkRuntime.ts: 6,039 -> 5,879 LOC (-160), drops 7 helper definitions
    + 4 type definitions, gains 2 import blocks

The visual helpers are pure (string in / string or props out, no React,
no Ink, no git). The types are structural; LogInkRuntime['ink'] still
types Ink's surface without bundling the ESM-only module.

No behavior change. tsc clean, lint clean, 680/680 suites & 6571/6571
tests passing.
